requirements.txt
Django
Pillow
requests
django-ckeditor